refactor: modularize CombatReplayModal component
- Extract utilities to combatReplayUtils.ts (195 LOC)
- Extract types to combatReplayTypes.ts (102 LOC)
- Extract SubEventCard component (109 LOC)
- Extract ActionCard component (78 LOC)
- Extract TurnCard component (154 LOC)
- Extract SyncLogPanel component (133 LOC)
- Add barrel export index.ts (47 LOC)
- Reduce main component from 1,151 LOC to 458 LOC (60% reduction)

All tests passing. Build successful.
